update(test-result): 합불 공개 조회 서비스 클래스명 변경
